Implement full Ivy support (and remove grizzl)
This commit enables the possibility to add custom actions for Ivy. Also it
removes the grizzl (deprecated package) features.

First it sets the value of `this-command` to 'fasd-findfile as
`ivy-completing-read assumes `this-command` to be its
caller (and the `read-from-minibuffer` for some reason modifies its value).

Second it defines a fasd-find-file-action which is set as ivy's default action,
or it will be evaluated after `completing-read` only if `ivy` is not available.
This prevents this part to get evaluated after some custom ivy-action was
selected.

An example custom action can be found in Spacemacs its fasd layer, which add an
action to (rip)grep through all files of some selected directory.